move iceberg test to separate file
